py-client/requirements.txt
web3>=6.0.0
sapphire-py>=0.2.0
python-dotenv>=1.0.0
